--- packages/scrabble-solver/src/components/PlainTiles/lib/index.ts
@@ -1,6 +1,0 @@
-export { createPlainTile } from './createPlainTile';
-export { createPlainTiles } from './createPlainTiles';
-export { getViewbox } from './getViewbox';
-export { getX } from './getX';
-export { getY } from './getY';
-export { randomize } from './randomize';
