--- packages/scrabble-solver/src/components/PlainTiles/lib/randomize.ts
@@ -1,1 +1,0 @@
-export const randomize = (value: number, maxChange: number): number => value + maxChange * 2 * (0.5 - Math.random());
